Resolved issue for Migration and created DB

+++ MedAppointment.Entities/Composition/OrganizationUserEntity.cs
@@ -2,8 +2,8 @@
 {
     public class OrganizationUserEntity : BaseEntity
     {
-        public int UserId { get; set; }
-        public int OrganizationId { get; set; }
+        public long UserId { get; set; }
+        public long OrganizationId { get; set; }
         public bool IsAdmin { get; set; }
 
         public OrganizationEntity? Organization { get; set; }
